+++ api/src/main/java/com/hcs/annotation/EnableMockMvc.java
@@ -1,11 +1,15 @@
-package com.hcs.config;
+package com.hcs.annotation;
 
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 import org.springframework.web.filter.CharacterEncodingFilter;
 
-import java.lang.annotation.*;
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
 /**
  * @AutoConfigureMockMvc : Mock 테스트 시 필요한 의존성을 제공해줌.
